Refactor code related to chat completions
