+++ NEWS.md
@@ -23,6 +23,7 @@ Corrections
  * When browsing your own shares, viewing recently shared downloads no longer requires a restart
  * Attempting to download files of disconnected users now displays the 'User logged off' status immediately
  * Column widths of the currently selected user browse tab are now saved
+ * Unified chat completion behavior of chat rooms and private chats
  * UI customizations are once again applied to the preferences dialog
  * Corrected the behavior of 'Abort User's Uploads' button in the uploads view
  * Text-To-Speech messages no longer overlap each other
