+++ pynicotine/config.py
@@ -160,8 +160,8 @@ class Config:
                 "replacewords": False,
                 "tab": True,
                 "cycle": False,
-                "dropdown": True,
-                "characters": 2,
+                "dropdown": False,
+                "characters": 3,
                 "roomnames": True,
                 "buddies": True,
                 "roomusers": True,
