+++ pynicotine/gtkgui/privatechat.py
@@ -27,22 +27,17 @@ from collections import deque
 from time import altzone
 from time import daylight
 
-from gi.repository import Gdk
 from gi.repository import Gio
 from gi.repository import GLib
 from gi.repository import Gtk
 
 from pynicotine import slskmessages
 from pynicotine.config import config
-from pynicotine.gtkgui.chatrooms import get_completion
 from pynicotine.gtkgui.utils import append_line
 from pynicotine.gtkgui.utils import auto_replace
 from pynicotine.gtkgui.utils import censor_chat
 from pynicotine.gtkgui.utils import copy_all_text
 from pynicotine.gtkgui.utils import delete_log
-from pynicotine.gtkgui.utils import entry_completion_find_match
-from pynicotine.gtkgui.utils import entry_completion_found_match
-from pynicotine.gtkgui.utils import keyval_to_hardware_keycode
 from pynicotine.gtkgui.utils import load_ui_elements
 from pynicotine.gtkgui.utils import open_log
 from pynicotine.gtkgui.utils import scroll_bottom
@@ -50,31 +45,30 @@ from pynicotine.gtkgui.utils import triggers_context_menu
 from pynicotine.gtkgui.widgets.iconnotebook import IconNotebook
 from pynicotine.gtkgui.widgets.popupmenu import PopupMenu
 from pynicotine.gtkgui.widgets.messagedialogs import option_dialog
+from pynicotine.gtkgui.widgets.textentry import ChatEntry
 from pynicotine.gtkgui.widgets.textentry import TextSearchBar
 from pynicotine.gtkgui.widgets.theme import get_user_status_color
 from pynicotine.gtkgui.widgets.theme import update_tag_visuals
 from pynicotine.gtkgui.widgets.theme import update_widget_visuals
 from pynicotine.logfacility import log
-from pynicotine.utils import add_alias
-from pynicotine.utils import expand_alias
+from pynicotine.utils import get_completion_list
 from pynicotine.utils import get_path
-from pynicotine.utils import is_alias
-from pynicotine.utils import unalias
 from pynicotine.utils import version
 
 
 CTCP_VERSION = "\x01VERSION\x01"
 
+# List of allowed commands. The implementation for them is in the ChatEntry class.
+CMDS = [
+    "/al ", "/alias ", "/un ", "/unalias ", "/w ", "/whois ", "/browse ", "/b ", "/ip ", "/pm ", "/m ", "/msg ",
+    "/s ", "/search ", "/us ", "/usearch ", "/rs ", "/rsearch ", "/bs ", "/bsearch ", "/j ", "/join ",
+    "/ad ", "/add ", "/buddy ", "/rem ", "/unbuddy ", "/ban ", "/ignore ", "/ignoreip ", "/unban ",
+    "/unignore ", "/clear ", "/cl ", "/a ", "/away ", "/q ", "/quit ", "/exit ", "/now ", "/rescan ",
+    "/info ", "/toggle ", "/ctcpversion "
+]
+
 
 class PrivateChats(IconNotebook):
-
-    CMDS = set(
-        [
-            "/alias ", "/unalias ", "/whois ", "/browse ", "/ip ", "/pm ", "/msg ", "/search ", "/usearch ", "/rsearch ",
-            "/bsearch ", "/join ", "/add ", "/buddy ", "/rem ", "/unbuddy ", "/ban ", "/ignore ", "/ignoreip ", "/unban ", "/unignore ",
-            "/clear ", "/quit ", "/exit ", "/rescan ", "/info ", "/ctcpversion "
-        ]
-    )
 
     def __init__(self, frame):
 
@@ -95,7 +89,7 @@ class PrivateChats(IconNotebook):
 
         self.connected = True
         self.users = {}
-        self.clist = []
+        self.completion_list = []
         self.private_message_queue = {}
 
         self.notebook.connect("switch-page", self.on_switch_chat)
@@ -270,9 +264,6 @@ class PrivateChats(IconNotebook):
 
         del self.users[tab.user]
 
-        # Update completions on exit
-        self.update_completions()
-
         if tab.user in config.sections["privatechat"]["users"]:
             config.sections["privatechat"]["users"].remove(tab.user)
 
@@ -307,22 +298,10 @@ class PrivateChats(IconNotebook):
 
     def update_completions(self):
 
-        config_words = config.sections["words"]
-        clist = [config.sections["server"]["login"], "nicotine"] + list(self.users.keys())
+        self.completion_list = get_completion_list(CMDS, self.frame.chatrooms.roomlist.server_rooms)
 
-        if config_words["buddies"]:
-            clist += [i[0] for i in config.sections["server"]["userlist"]]
-
-        if config_words["aliases"]:
-            clist += ["/" + k for k in list(config.sections["server"]["command_aliases"].keys())]
-
-        if config_words["commands"]:
-            clist += self.CMDS
-
-        self.clist = clist
-
-        for user in list(self.users.values()):
-            user.get_completion_list(clist=list(self.clist))
+        for user in self.users.values():
+            user.set_completion_list(self.completion_list)
 
 
 class PrivateChat:
@@ -338,26 +317,12 @@ class PrivateChat:
         self.autoreplied = False
         self.offlinemessage = False
         self.status = status
-        self.clist = []
 
         # Text Search
         TextSearchBar(self.ChatScroll, self.SearchBar, self.SearchEntry)
 
-        # Spell Check
-        if self.frame.spell_checker is None:
-            self.frame.init_spell_checker()
-
-        if self.frame.spell_checker and config.sections["ui"]["spellcheck"]:
-            from gi.repository import Gspell
-            spell_buffer = Gspell.EntryBuffer.get_from_gtk_entry_buffer(self.ChatLine.get_buffer())
-            spell_buffer.set_spell_checker(self.frame.spell_checker)
-            spell_view = Gspell.Entry.get_from_gtk_entry(self.ChatLine)
-            spell_view.set_inline_spell_checking(True)
-
-        self.ChatCompletion.set_model(Gtk.ListStore(str))
-        self.ChatCompletion.set_text_column(0)
-        self.ChatCompletion.set_match_func(entry_completion_find_match)
-        self.ChatCompletion.connect("match-selected", entry_completion_found_match)
+        # Chat Entry
+        self.entry = ChatEntry(self.frame, self.ChatLine, user, slskmessages.MessageUser, self.send_message, CMDS, self.ChatScroll)
 
         self.Log.set_active(config.sections["logging"]["privatechat"])
 
@@ -388,7 +353,6 @@ class PrivateChat:
         self.create_tags()
         self.update_visuals()
 
-        self.chats.update_completions()
         self.read_private_log()
 
     def read_private_log(self):
@@ -552,6 +516,9 @@ class PrivateChat:
 
     def send_message(self, text, bytestring=False):
 
+        if not self.chats.connected:
+            return
+
         user_text = self.frame.np.pluginhandler.outgoing_private_chat_event(self.user, text)
         if user_text is None:
             return
@@ -593,183 +560,6 @@ class PrivateChat:
             self.frame.np.queue.append(slskmessages.MessageUser(self.user, payload))
 
         self.frame.np.pluginhandler.outgoing_private_chat_notification(self.user, text)
-
-    def on_enter(self, widget):
-
-        text = widget.get_text()
-
-        if not text:
-            widget.set_text("")
-            return
-
-        if is_alias(text):
-            new_text = expand_alias(text)
-
-            if not new_text:
-                log.add(_('Alias "%s" returned nothing'), text)
-                return
-
-            if new_text[:2] == "//":
-                new_text = new_text[1:]
-
-            self.frame.np.queue.append(slskmessages.MessageUser(self.user, auto_replace(new_text)))
-            widget.set_text("")
-            return
-
-        s = text.split(" ", 1)
-        cmd = s[0]
-        if len(s) == 2 and s[1]:
-            args = arg_self = s[1]
-        else:
-            args = ""
-            arg_self = self.user
-
-        if cmd in ("/alias", "/al"):
-
-            append_line(self.ChatScroll, add_alias(args), None, "")
-            if config.sections["words"]["aliases"]:
-                self.frame.chatrooms.update_completions()
-                self.frame.privatechats.update_completions()
-
-        elif cmd in ("/unalias", "/un"):
-
-            append_line(self.ChatScroll, unalias(args), None, "")
-            if config.sections["words"]["aliases"]:
-                self.frame.chatrooms.update_completions()
-                self.frame.privatechats.update_completions()
-
-        elif cmd in ("/join", "/j"):
-            if args:
-                self.frame.np.queue.append(slskmessages.JoinRoom(args))
-
-        elif cmd in ("/w", "/whois", "/info"):
-            self.frame.local_user_info_request(arg_self)
-            self.frame.change_main_page("userinfo")
-
-        elif cmd in ("/b", "/browse"):
-            self.frame.browse_user(arg_self)
-            self.frame.change_main_page("userbrowse")
-
-        elif cmd == "/ip":
-            self.frame.np.ip_requested.add(arg_self)
-            self.frame.np.queue.append(slskmessages.GetPeerAddress(arg_self))
-
-        elif cmd == "/pm":
-            if args:
-                self.frame.privatechats.send_message(args, show_user=True)
-                self.frame.change_main_page("private")
-
-        elif cmd in ("/m", "/msg"):
-            if args:
-                s = args.split(" ", 1)
-                user = s[0]
-                if len(s) == 2:
-                    msg = s[1]
-                else:
-                    msg = None
-
-                self.frame.privatechats.send_message(user, msg, show_user=True)
-                self.frame.change_main_page("private")
-
-        elif cmd in ("/s", "/search"):
-            if args:
-                self.frame.searches.do_search(args, 0)
-                self.frame.on_search(None)
-                self.frame.change_main_page("search")
-
-        elif cmd in ("/us", "/usearch"):
-            if args:
-                self.frame.searches.do_search(args, 3, [self.user])
-                self.frame.on_search(None)
-                self.frame.change_main_page("search")
-
-        elif cmd in ("/rs", "/rsearch"):
-            if args:
-                self.frame.searches.do_search(args, 1)
-                self.frame.on_search(None)
-                self.frame.change_main_page("search")
-
-        elif cmd in ("/bs", "/bsearch"):
-            if args:
-                self.frame.searches.do_search(args, 2)
-                self.frame.on_search(None)
-                self.frame.change_main_page("search")
-
-        elif cmd in ("/ad", "/add", "/buddy"):
-            if args:
-                self.frame.userlist.add_to_list(args)
-
-        elif cmd in ("/rem", "/unbuddy"):
-            if args:
-                self.frame.userlist.remove_from_list(args)
-
-        elif cmd == "/ban":
-            if args:
-                self.frame.np.network_filter.ban_user(args)
-
-        elif cmd == "/ignore":
-            if args:
-                self.frame.np.network_filter.ignore_user(args)
-
-        elif cmd == "/ignoreip":
-            if args:
-                self.frame.np.network_filter.ignore_ip(args)
-
-        elif cmd == "/unban":
-            if args:
-                self.frame.np.network_filter.unban_user(args)
-
-        elif cmd == "/unignore":
-            if args:
-                self.frame.np.network_filter.unignore_user(args)
-
-        elif cmd == "/ctcpversion":
-            if arg_self:
-                self.frame.privatechats.send_message(arg_self, CTCP_VERSION, show_user=True, bytestring=True)
-
-        elif cmd in ("/clear", "/cl"):
-            self.ChatScroll.get_buffer().set_text("")
-
-        elif cmd in ("/a", "/away"):
-            self.frame.on_away(None)
-
-        elif cmd in ("/q", "/quit", "/exit"):
-            self.frame.on_quit(None)
-            return
-
-        elif cmd in ("/c", "/close"):
-            self.on_close(None)
-
-        elif cmd == "/now":
-            self.display_now_playing()
-
-        elif cmd == "/rescan":
-            # Rescan public shares if needed
-            if not config.sections["transfers"]["friendsonly"] and config.sections["transfers"]["shared"]:
-                self.frame.on_rescan()
-
-            # Rescan buddy shares if needed
-            if config.sections["transfers"]["enablebuddyshares"]:
-                self.frame.on_buddy_rescan()
-
-        elif cmd[:1] == "/" and self.frame.np.pluginhandler.trigger_private_command_event(self.user, cmd[1:], args):
-            pass
-
-        elif cmd[:1] == "/" and cmd != "/me" and cmd[:2] != "//":
-            log.add(_("Command %s is not recognized"), text)
-            return
-
-        else:
-            if text[:2] == "//":
-                text = text[1:]
-
-            if self.chats.connected:
-                self.send_message(text)
-                widget.set_text("")
-
-            return
-
-        widget.set_text("")
 
     def display_now_playing(self):
         self.frame.np.now_playing.display_now_playing(callback=self.send_message)
@@ -833,69 +623,11 @@ class PrivateChat:
         color = get_user_status_color(self.status)
         update_tag_visuals(self.tag_username, color)
 
-    def get_completion_list(self, ix=0, text="", clist=None):
-
-        config_words = config.sections["words"]
-
-        completion = self.ChatLine.get_completion()
-        completion.set_popup_single_match(not config_words["onematch"])
-        completion.set_minimum_key_length(config_words["characters"])
-
-        liststore = completion.get_model()
-        liststore.clear()
-
-        if clist is None:
-            clist = []
-
-        if not config_words["tab"]:
-            return
-
-        # no duplicates
-        def _combilower(x):
-            try:
-                return str.lower(x)
-            except Exception:
-                return str.lower(x)
-
-        clist = list(set(clist))
-        clist.sort(key=_combilower)
-
-        completion.set_popup_completion(False)
-
-        if config_words["dropdown"]:
-            for word in clist:
-                liststore.append([word])
-
-            completion.set_popup_completion(True)
-
-        self.clist = clist
-
-    def on_key_press(self, widget, event):
-
-        keycode = event.hardware_keycode
-
-        if keycode not in keyval_to_hardware_keycode(Gdk.KEY_Tab):
-            return False
-
-        if not config.sections["words"]["tab"]:
-            return False
-
-        ix = widget.get_position()
-        text = widget.get_text()[:ix].split(" ")[-1]
-        preix = ix - len(text)
-
-        completion, single = get_completion(text, self.chats.clist)
-
-        if completion:
-            widget.delete_text(preix, ix)
-            widget.insert_text(completion, ix)
-            widget.set_position(preix + len(completion))
-
-        widget.stop_emission_by_name("key_press_event")
-        return True
-
     def on_close(self, *args):
         self.chats.remove_tab(self)
 
     def on_close_all_tabs(self, *args):
         self.chats.remove_all_pages()
+
+    def set_completion_list(self, completion_list):
+        self.entry.set_completion_list(completion_list)
