+++ pynicotine/gtkgui/ui/chatrooms.ui
@@ -132,9 +132,6 @@
                             <property name="has-focus">1</property>
                             <property name="hexpand">1</property>
                             <property name="width-chars">5</property>
-                            <property name="completion">ChatCompletion</property>
-                            <signal name="activate" handler="on_enter" swapped="no"/>
-                            <signal name="key-press-event" handler="on_key_press" swapped="no"/>
                           </object>
                         </child>
                         <child>
@@ -313,8 +310,5 @@
         </child>
       </object>
     </child>
-  </object>
-  <object class="GtkEntryCompletion" id="ChatCompletion">
-    <property name="inline-completion">1</property>
   </object>
 </interface>
