+++ pynicotine/gtkgui/ui/privatechat.ui
@@ -64,9 +64,6 @@
                 <property name="visible">1</property>
                 <property name="can-focus">1</property>
                 <property name="hexpand">1</property>
-                <property name="completion">ChatCompletion</property>
-                <signal name="activate" handler="on_enter" swapped="no"/>
-                <signal name="key-press-event" handler="on_key_press" swapped="no"/>
               </object>
             </child>
             <child>
@@ -107,8 +104,5 @@
         </child>
       </object>
     </child>
-  </object>
-  <object class="GtkEntryCompletion" id="ChatCompletion">
-    <property name="inline-completion">1</property>
   </object>
 </interface>
